+++ src/main/java/com/ddukbbegi/api/auth/service/AuthServiceImpl.java
@@ -8,8 +8,7 @@ import com.ddukbbegi.api.auth.dto.response.SignupResponseDto;
 import com.ddukbbegi.api.user.entity.User;
 import com.ddukbbegi.api.user.enums.UserRole;
 import com.ddukbbegi.api.user.repository.UserRepository;
-import com.ddukbbegi.common.component.ResultCode;
-import com.ddukbbegi.common.auth.JwtUtil;
+import com.ddukbbegi.common.jwt.JwtUtil;
 import com.ddukbbegi.common.config.PasswordEncoder;
 import com.ddukbbegi.common.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
