+++ src/main/java/com/ddukbbegi/api/user/controller/UserController.java
@@ -5,8 +5,8 @@ import com.ddukbbegi.api.user.dto.request.*;
 import com.ddukbbegi.api.user.dto.response.MyInfoResponseDto;
 import com.ddukbbegi.api.user.dto.response.UserInfoResponseDto;
 import com.ddukbbegi.api.user.service.UserService;
-import com.ddukbbegi.common.auth.CustomUserDetails;
-import com.ddukbbegi.common.auth.JwtUtil;
+import com.ddukbbegi.api.user.entity.CustomUserDetails;
+import com.ddukbbegi.common.jwt.JwtUtil;
 import com.ddukbbegi.common.component.BaseResponse;
 import com.ddukbbegi.common.component.ResultCode;
 import jakarta.validation.Valid;
