+++ src/main/java/com/ddukbbegi/common/config/AuditorAwareImpl.java
@@ -1,6 +1,6 @@
 package com.ddukbbegi.common.config;
 
-import com.ddukbbegi.common.auth.CustomUserDetails;
+import com.ddukbbegi.api.user.entity.CustomUserDetails;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
